feat: versioning for drive's withdrawal functions
