+++ packages/rs-drive-abci/src/execution/platform_events/withdrawals/mod.rs
@@ -1,6 +1,6 @@
 pub(in crate::execution) mod append_signatures_and_broadcast_withdrawal_transactions;
 pub(in crate::execution) mod build_untied_withdrawal_transactions_from_documents;
 pub(in crate::execution) mod dequeue_and_build_unsigned_withdrawal_transactions;
-mod fetch_transactions_block_inclusion_status;
+pub(in crate::execution) mod fetch_transactions_block_inclusion_status;
 pub(in crate::execution) mod pool_withdrawals_into_transactions_queue;
 pub(in crate::execution) mod update_broadcasted_withdrawal_statuses;
