+++ packages/rs-drive/src/drive/identity/withdrawals/mod.rs
@@ -1,12 +1,10 @@
-/// Functions related to updating of a withdrawal status
-pub mod documents;
+/// Functions related to withdrawal documents
+pub mod document;
+
 /// Functions and constants related to GroveDB paths
 pub mod paths;
-/// Functions related to withdrawal queue
-// TODO: Rename to transaction_queue
-pub mod queue;
-/// Functions related to transaction index counter
-pub mod transaction_index;
+/// Functions related to withdrawal transactions
+pub mod transaction;
 
 /// Simple type alias for withdrawal transaction with it's index
 pub type WithdrawalTransactionIndexAndBytes = (WithdrawalTransactionIndex, Vec<u8>);
