+++ packages/rs-drive/src/drive/identity/withdrawals/transaction/mod.rs
@@ -1,0 +1,5 @@
+/// Functions related to withdrawal queue
+pub mod queue;
+
+/// Functions related to transaction index counter
+pub mod index;
